requirements.txt
python-dotenv
chromadb 
sentence-transformers 
unstructured[pdf]
google.generativeai